+++ src/components/content/medications/patient-medications-outside.tsx
@@ -1,12 +1,13 @@
 import { useAddMedicationForm } from "./helpers/add-new-med-drawer";
 import { medicationFilters } from "./helpers/filters";
 import { PatientMedicationsBase } from "./helpers/patient-medications-base";
+import { useToggleArchive } from "../hooks/use-toggle-archive";
 import { withErrorBoundary } from "@/components/core/error-boundary";
 import { RowActionsProps } from "@/components/core/table/table";
-import { toggleArchive } from "@/fhir/basic";
 import { MedicationStatementModel } from "@/fhir/models";
 import { useQueryAllPatientMedications } from "@/hooks/use-medications";
-import { useCTW } from "@/index";
+import { useBaseTranslations } from "@/i18n";
+import { Spinner } from "@/index";
 import { QUERY_KEY_OTHER_PROVIDER_MEDICATIONS } from "@/utils/query-keys";
 
 export type PatientMedicationsOutsideProps = {
@@ -52,21 +53,29 @@ const getRowActions =
 type RowActionsProps2 = RowActionsProps<MedicationStatementModel> & ExtraRowActionProps;
 
 const RowActions = ({ record, onAddToRecord }: RowActionsProps2) => {
+  const { t } = useBaseTranslations();
   const showAddMedicationForm = useAddMedicationForm();
-  const { getRequestContext } = useCTW();
+  const { isLoading, toggleArchive } = useToggleArchive(
+    record,
+    QUERY_KEY_OTHER_PROVIDER_MEDICATIONS
+  );
+  const archiveLabel = record.isArchived ? t("resourceTable.restore") : t("resourceTable.dismiss");
 
   return (
     <div className="ctw-flex ctw-space-x-2">
       <button
         type="button"
         className="ctw-btn-default"
-        onClick={async () => {
-          await toggleArchive(record, await getRequestContext(), [
-            QUERY_KEY_OTHER_PROVIDER_MEDICATIONS,
-          ]);
-        }}
+        disabled={isLoading}
+        onClick={toggleArchive}
       >
-        {record.isArchived ? "Restore" : "Dismiss"}
+        {isLoading ? (
+          <div className="ctw-flex">
+            <Spinner className="ctw-mx-4 ctw-align-middle" />
+          </div>
+        ) : (
+          archiveLabel
+        )}
       </button>
 
       <button
@@ -74,6 +83,7 @@ const RowActions = ({ record, onAddToRecord }: RowActionsProps2) => {
         className="ctw-btn-primary ctw-ml-1 ctw-capitalize"
         data-zus-telemetry-click="Add to record"
         data-testid="add-to-record"
+        disabled={isLoading}
         onClick={() => {
           if (onAddToRecord) {
             onAddToRecord(record);
@@ -82,7 +92,7 @@ const RowActions = ({ record, onAddToRecord }: RowActionsProps2) => {
           }
         }}
       >
-        Add
+        {t("resourceTable.add")}
       </button>
     </div>
   );