+++ src/components/content/patient-history/use-patient-history.tsx
@@ -1,3 +1,4 @@
+import { format } from "date-fns";
 import { useEffect, useState } from "react";
 import { PatientHistoryStatus } from "./patient-history-message-status";
 import { PatientHistoryRequestDrawer } from "../patient-history-request-drawer";
@@ -90,6 +91,7 @@ export type GetBuilderRefreshHistoryMessagesParams = {
   offset?: number;
   patientId?: string;
   status?: string;
+  excludeFutureJobs?: boolean;
 };
 
 export async function getBuilderRefreshHistoryMessages({
@@ -98,6 +100,7 @@ export async function getBuilderRefreshHistoryMessages({
   offset = 0,
   patientId,
   status,
+  excludeFutureJobs,
 }: GetBuilderRefreshHistoryMessagesParams) {
   const baseUrl = new URL(`${getZusApiBaseUrl(requestContext.env)}/patient-history/jobs?`);
 
@@ -110,6 +113,7 @@ export async function getBuilderRefreshHistoryMessages({
         : "",
       "filter[patient-id]": patientId ? `${patientId}` : "",
       "filter[status]": status ? `${status}` : "",
+      ...(!!excludeFutureJobs && { "filter[targetDate][until]": format(Date.now(), "yyyy-MM-dd") }),
     },
     (value) => !value
   );