+++ src/fhir/models/patient-history.ts
@@ -19,12 +19,19 @@ export class PatientHistoryRequestModel {
     return this.historyInfo?.attributes.providers;
   }
 
+  get lastUpdatedAt() {
+    return format(
+      new Date(Number(this.historyInfo?.attributes.lastUpdatedAt) * 1000),
+      "M/d/yy h:mm a"
+    );
+  }
+
   get key() {
     return this.historyInfo?.id || "";
   }
 
-  get lastRetrievedAt() {
-    return formatISODateStringToDate(this.historyInfo?.attributes.createdAt);
+  get targetDate() {
+    return formatISODateStringToDate(this.historyInfo?.attributes.targetDate);
   }
 
   get createdAt() {